// photo-app/src/API.ts
export type Post = {
  __typename: "Post",
  id: string,
  title: string,
  imageKey?: string | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type OnCreatePostSubscription = {
  onCreatePost?:  {
    __typename: "Post",
    id: string,
    title: string,
    imageKey?: string | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};